requirements.txt
Flask==3.0.2
flask-sqlalchemy==3.1.1
flask-restx==1.3.0
psycopg2-binary==2.9.9
pandas==2.2.1
numpy==1.26.4
scipy==1.12.0
python-dotenv==1.0.1
marshmallow==3.20.2
scikit-learn==1.6.1
gunicorn==21.2.0 